requirements.txt
django == 4.2.9
django_extensions == 3.2.3
djangorestframework == 3.14.0
pytest == 7.4.4
coverage == 7.4.0
pytest-django == 4.7.0
pytest-xdist == 3.5.0
pytest-cov == 4.1.0
django-cors-headers == 4.3.1
pillow == 10.2.0
django-jazzmin == 2.6.0